Issue 56 57 60 64 67 (#71)
* (maint) Fix typo in code

* Issue 56: Rewrite SQL query for doubloons

* (maint) Reorder code

* Issue 57: Add deletion confirmation

* (maint) Fix NLog configuration

* Issue 64: Add validation on alias creation

* Issue 67: Fix crash when no statistics

* Issue 60: Close settings on pressing ESC

+++ src/Lanceur/Views/DoubloonsViewModel.cs
@@ -10,6 +10,9 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reactive;
 using System.Reactive.Concurrency;
+using System.Reactive.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Lanceur.Views
 {
@@ -17,6 +20,7 @@ namespace Lanceur.Views
     {
         #region Fields
 
+        private readonly Interaction<string, bool> _confirmRemove;
         private readonly IDataService _service;
         private readonly IThumbnailManager _thumbnailManager;
 
@@ -35,7 +39,7 @@ namespace Lanceur.Views
             notify ??= l.GetService<IUserNotification>();
             _service = service ?? l.GetService<IDataService>();
             _thumbnailManager = thumbnailManager ?? l.GetService<IThumbnailManager>();
-
+            _confirmRemove = Interactions.YesNoQuestion(uiThread);
 
             uiThread ??= RxApp.MainThreadScheduler;
             poolThread ??= RxApp.TaskpoolScheduler;
@@ -43,7 +47,7 @@ namespace Lanceur.Views
             Activate = ReactiveCommand.Create(OnActivate, outputScheduler: uiThread);
             Activate.ThrownExceptions.Subscribe(ex => notify.Error(ex.Message, ex));
 
-            RemoveSelected = ReactiveCommand.Create(OnRemoveSelected, outputScheduler: uiThread);
+            RemoveSelected = ReactiveCommand.CreateFromTask(OnRemoveSelected, outputScheduler: uiThread);
             RemoveSelected.ThrownExceptions.Subscribe(ex => notify.Error(ex.Message, ex));
 
             this.WhenAnyObservable(vm => vm.Activate)
@@ -55,6 +59,7 @@ namespace Lanceur.Views
         #region Properties
 
         public ReactiveCommand<Unit, ObservableCollection<SelectableAliasQueryResult>> Activate { get; }
+        public Interaction<string, bool> ConfirmRemove => _confirmRemove;
         [Reactive] public ObservableCollection<SelectableAliasQueryResult> Doubloons { get; set; }
         public ReactiveCommand<Unit, Unit> RemoveSelected { get; }
 
@@ -70,14 +75,20 @@ namespace Lanceur.Views
             return results;
         }
 
-        private void OnRemoveSelected()
+        private async Task OnRemoveSelected()
         {
             var toDel = (from d in Doubloons
                          where d.IsSelected
                          select d).ToList();
-            foreach (var item in toDel) { Doubloons.Remove(item); }
-            _service.Remove(toDel);
-            Toast.Information($"Removed {toDel.Count} alias(es).");
+
+            var count = toDel?.Count() ?? 0;
+            var remove = await _confirmRemove.Handle($"{count}");
+            if (remove && count > 0)
+            {
+                foreach (var item in toDel) { Doubloons.Remove(item); }
+                _service.Remove(toDel);
+                Toast.Information($"Removed {toDel.Count} alias(es).");
+            }
         }
 
         #endregion Methods
