+++ src/Lanceur/Views/InvalidAliasViewModel.cs
@@ -9,6 +9,8 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reactive;
 using System.Reactive.Concurrency;
+using System.Reactive.Linq;
+using System.Threading.Tasks;
 
 namespace Lanceur.Views
 {
@@ -16,6 +18,7 @@ namespace Lanceur.Views
     {
         #region Fields
 
+        private readonly Interaction<string, bool> _confirmRemove;
         private readonly IDataService _service;
 
         #endregion Fields
@@ -34,11 +37,12 @@ namespace Lanceur.Views
             var l = Locator.Current;
             notify ??= l.GetService<IUserNotification>();
             _service = service ?? l.GetService<IDataService>();
+            _confirmRemove = Interactions.YesNoQuestion(uiThread);
 
             Activate = ReactiveCommand.Create(OnActivate, outputScheduler: uiThread);
             Activate.ThrownExceptions.Subscribe(ex => notify.Error(ex.Message, ex));
 
-            RemoveSelected = ReactiveCommand.Create(OnRemoveSelected, outputScheduler: uiThread);
+            RemoveSelected = ReactiveCommand.CreateFromTask(OnRemoveSelected, outputScheduler: uiThread);
             RemoveSelected.ThrownExceptions.Subscribe(ex => notify.Error(ex.Message, ex));
 
             this.WhenAnyObservable(vm => vm.Activate)
@@ -51,6 +55,7 @@ namespace Lanceur.Views
 
         public ReactiveCommand<Unit, ObservableCollection<SelectableAliasQueryResult>> Activate { get; }
 
+        public Interaction<string, bool> ConfirmRemove => _confirmRemove;
         [Reactive] public ObservableCollection<SelectableAliasQueryResult> InvalidAliases { get; set; }
 
         public ReactiveCommand<Unit, Unit> RemoveSelected { get; }
@@ -65,14 +70,19 @@ namespace Lanceur.Views
             return new(results);
         }
 
-        private void OnRemoveSelected()
+        private async Task OnRemoveSelected()
         {
             var toDel = (from d in InvalidAliases
                          where d.IsSelected
                          select d).ToList();
-            foreach (var item in toDel) { InvalidAliases.Remove(item); }
-            _service.Remove(toDel);
-            Toast.Information($"Removed {toDel.Count} alias(es).");
+            var count = toDel?.Count() ?? 0;
+            var remove = await _confirmRemove.Handle($"{count}");
+            if (remove && count > 0)
+            {
+                foreach (var item in toDel) { InvalidAliases.Remove(item); }
+                _service.Remove(toDel);
+                Toast.Information($"Removed {toDel.Count} alias(es).");
+            }
         }
 
         #endregion Methods
